+++ js/constants.js
@@ -1,0 +1,6 @@
+const messageText = 'Всё отлично! В целом всё неплохо. Но не всё. Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально. Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше. Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше. Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!';
+const firstNames = ['Антон', 'Василий', 'Петр', 'Кирилл', 'Илья', 'Станислав', 'Максим'];
+const lastNames = ['Иванов', 'Сидоров', 'Петров', 'Кузнецов', 'Семенов', 'Джаваскриптов'];
+const photoDescriptions = ['Космические линии', 'Игра света и тени', 'Цветовые пятна', 'Энергия движения', 'Сложные узоры', 'Необычные формы', 'Гармония цвета', 'Размытые контуры', 'Природные текстуры', 'Загадочный силуэт'];
+
+export {messageText, firstNames, lastNames, photoDescriptions}
