+++ js/utils.js
@@ -1,0 +1,12 @@
+function getRandomInteger (a, b) {
+    const lower = Math.ceil(Math.min(a, b));
+    const upper = Math.floor(Math.max(a, b));
+    const result = Math.random() * (upper - lower + 1) + lower;
+    return Math.floor(result);
+}
+
+function getRandomItem (arr){
+    return arr[getRandomInteger(0,arr.length - 1)]
+}
+
+export {getRandomInteger, getRandomItem}
